refactor(sanity): move base components from form/inputs/DateInputs to ui-components
